requirements.txt
django==4.1.5
djangorestframework==3.14.0
drf-spectacular==0.25.1
google-cloud-logging==3.5.0
gunicorn==20.1.0
psutil==5.9.4
structlog==22.3.0
uvicorn==0.20.0
